No half for mnist

+++ test/models/test_mnist.py
@@ -8,14 +8,12 @@ from datasets import fetch_mnist
 
 # load the mnist dataset
 X_train, Y_train, X_test, Y_test = fetch_mnist()
-X_train = X_train.astype(np.float16)
-X_test = X_test.astype(np.float16)
 
 # create a model
 class TinyBobNet:
   def __init__(self):
-    self.l1 = Tensor.scaled_uniform(784, 128).half()
-    self.l2 = Tensor.scaled_uniform(128, 10).half()
+    self.l1 = Tensor.scaled_uniform(784, 128)
+    self.l2 = Tensor.scaled_uniform(128, 10)
 
   def parameters(self):
     return optim.get_parameters(self)
